fix(activity-check): missing intl prop
Refactor activity check component to a functional component to utilize
the `useIntl` hook. Also corvert it to Typescript.
