+++ bigbluebutton-html5/imports/ui/components/activity-check/container.jsx
@@ -19,7 +19,6 @@ const ActivityCheckContainer = () => {
   return (
     <ActivityCheck
       userActivitySign={userActivitySign}
-      inactivityCheck={inactivityWarningDisplay}
       responseDelay={inactivityWarningTimeoutSecs}
     />
   );
